Fix for missing psycopg2 package

+++ run_tests.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 initialize_reactions_db development.ini
 python setup.py behave_test
-rm reactions.sqlite
+if [ -f reactions.sqlite ]; then
+    rm reactions.sqlite
+fi
